// src/app/shared/loca-model/user-loca-dto.ts
export class UserLocaDto {
    company_name: string;
    firstName: string;
    id: string;
    isRoleDeleted: string;
    isUserDeleted: string;
    lastName: string;
    otp: string;
    password: string;
    phone: string;
    roleApprovedBy: number;
    roleId: number;
    roleName: string;
    token: string;
    userId: number;
    username: string;

    constructor() {

    }
}